output/lights: extract per-version lighting strategies
Split the lighting model into a strategy pattern with one
LIGHTING_MODEL per lighting family (TR1/2 scalar, TR3 smoothed RGB;
TR4 rides the TR3 model for now). The public Output_* API is
unchanged; the g_TRVersion branches inside lights.c and uniforms.c
move into per-family vtable hooks, including the LightSource UBO
uploads. No behavior change.
