+++ src/meson.build
@@ -648,7 +648,9 @@ common_sources = [
   'trx/game/output/common.c',
   'trx/game/output/draw.c',
   'trx/game/output/func.c',
-  'trx/game/output/lights.c',
+  'trx/game/output/lights/common.c',
+  'trx/game/output/lights/tr12.c',
+  'trx/game/output/lights/tr3.c',
   'trx/game/output/mesh_batcher/batcher.c',
   'trx/game/output/mesh_batcher/mesh.c',
   'trx/game/output/mesh_batcher/mesh_builder.c',
